% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZlibrary_palettes.R
\name{largest_dist_helper}
\alias{largest_dist_helper}
\title{If n colors in palette are fewer than total palette length, select those colors that maximize distance}
\usage{
largest_dist_helper(pal, n)
}
\arguments{
\item{pal}{color palette values.}

\item{n}{number of desired colors in palette; defaults to all colors in given palette.}
}
\value{
list of colors to keep in palette of n length
}
\description{
If n colors in palette are fewer than total palette length, select those colors that maximize distance
}
