% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZlibrary_palettes.R
\name{MVZ_cb}
\alias{MVZ_cb}
\title{Check whether palettes are colorblind friendly
Uses the \link[colorblindcheck]{palette_check} function}
\usage{
MVZ_cb(name)
}
\arguments{
\item{name}{name of MVZ library palette (e.g., \code{"NaturalHistory"})}
}
\value{
values for color vision deficiencies
}
\description{
Check whether palettes are colorblind friendly
Uses the \link[colorblindcheck]{palette_check} function
}
