% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZlibrary_palettes.R
\name{MVZ_palette}
\alias{MVZ_palette}
\title{Generates MVZ library palette}
\usage{
MVZ_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{name of palette. Options are: \code{"NaturalHistory"}, \code{"LifeHistories"}, \code{"theAuk"}, \code{"Zoology"}, \code{"FieldNotes"}, \code{"Ensatina"}, \code{"WesternBirds"}.}

\item{n}{number of desired colors in palette; defaults to all colors in given palette.}

\item{type}{either \code{"continuous"} or \code{"discrete"} (default). Using \code{"continuous"} will automatically interpolate between colors.}
}
\value{
a vector of colors
}
\description{
This function is largely adapted from the \code{wes_palette()} function in the \href{https://github.com/karthik/wesanderson/tree/master}{wesanderson} package.
}
\examples{
MVZ_palette(name = "NaturalHistory", type = "continuous", n = 15)
MVZ_palette(name = "theAuk", n = 3)

}
