% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZlibrary_palettes.R
\docType{data}
\name{MVZ_palettes}
\alias{MVZ_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
MVZ_palettes
}
\description{
Use \code{\link{MVZ_palette}} to construct palettes of desired length.
}
\keyword{datasets}
