% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZlibrary_palettes.R
\name{MVZ_type}
\alias{MVZ_type}
\title{List whether palette is qualitative, diverging, or monochromatic}
\usage{
MVZ_type(name = "all")
}
\arguments{
\item{name}{list type of given palette (e.g., \code{"NaturalHistory"}), list all palettes of desired type (e.g., \code{"monochromatic"}), or list types of all (\code{"all"}; default)}
}
\value{
prints types of palettes
}
\description{
List whether palette is qualitative, diverging, or monochromatic
}
\examples{
MVZ_type(name = "NaturalHistory")

}
